+++ slackclient/slackrequest.py
@@ -3,6 +3,7 @@ import platform
 import sys
 
 import requests
+import json
 import six
 
 from .version import __version__
@@ -60,26 +61,48 @@ class SlackRequest(object):
                 than slack.com
         """
 
+        url = 'https://{0}/api/{1}'.format(domain, request)
+
+        # Override token header if `token` is passed in post_data
+        if post_data is not None and "token" in post_data:
+            token = post_data['token']
+
+        # Set user-agent and auth headers
+        headers = {
+            'user-agent': self.get_user_agent(),
+            'Authorization': 'Bearer {}'.format(token)
+        }
+
         # Pull file out so it isn't JSON encoded like normal fields.
         # Only do this for requests that are UPLOADING files; downloading files
         # use the 'file' argument to point to a File ID.
         post_data = post_data or {}
+
+        # Move singular file objects into `files`
         upload_requests = ['files.upload']
+
+        # Move file content into requests' `files` param
         files = None
         if request in upload_requests:
             files = {'file': post_data.pop('file')} if 'file' in post_data else None
 
+        # Check for plural fields and convert them to comma-separated strings if needed
+        for field in {'channels', 'users', 'types'} & set(post_data.keys()):
+            if isinstance(post_data[field], list):
+                post_data[field] = ",".join(post_data[field])
+
+        # Convert any params which are list-like to JSON strings
+        # Example: `attachments` is a dict, and needs to be passed as JSON
         for k, v in six.iteritems(post_data):
-            if not isinstance(v, six.string_types):
+            if isinstance(v, (list, dict)):
                 post_data[k] = json.dumps(v)
 
-        url = 'https://{0}/api/{1}'.format(domain, request)
-        post_data['token'] = token
-        headers = {'user-agent': self.get_user_agent()}
-
-        return requests.post(url,
-                             headers=headers,
-                             data=post_data,
-                             files=files,
-                             timeout=timeout,
-                             proxies=self.proxies)
+        # Submit the request
+        return requests.post(
+            url,
+            headers=headers,
+            data=post_data,
+            files=files,
+            timeout=timeout,
+            proxies=self.proxies
+        )