+++ tests/test_server.py
@@ -1,10 +1,14 @@
 import json
+import time
 import pytest
 import requests
+import responses
+from mock import patch
 from slackclient.user import User
-from slackclient.server import Server, SlackLoginError
+from slackclient.server import Server, SlackLoginError, SlackConnectionError
 from slackclient.channel import Channel
 from requests.packages.urllib3.exceptions import InsecureRequestWarning
+
 
 requests.packages.urllib3.disable_warnings(InsecureRequestWarning)
 
@@ -16,14 +20,60 @@ def rtm_start_fixture():
     return json_login_data
 
 
-def test_server(server):
+def test_server():
+    server = Server("valid_token", connect=False)
     assert type(server) == Server
+
+    # The server eqs to a string, either the token or workspace domain
+    assert server == "valid_token"
+    assert server != "invalid_token"
+
+
+def test_server_connect(rtm_start_fixture):
+    with responses.RequestsMock() as rsps:
+        rsps.add(
+            responses.POST,
+            "https://slack.com/api/rtm.start",
+            status=200,
+            json=rtm_start_fixture
+        )
+
+        Server("token", connect=True)
+
+        for call in rsps.calls:
+            assert call.request.url in [
+                "https://slack.com/api/rtm.start"
+            ]
 
 
 def test_server_is_hashable(server):
     server_map = {server: server.token}
     assert server_map[server] == 'xoxp-1234123412341234-12341234-1234'
     assert (server_map[server] == 'foo') is False
+
+
+@patch('time.sleep', return_value=None)
+def test_rate_limiting(patched_time_sleep, server):
+    # Testing for rate limit retry headers
+    with responses.RequestsMock() as rsps:
+        rsps.add(
+            responses.POST,
+            "https://slack.com/api/rtm.start",
+            status=429,
+            json={"ok": False},
+            headers={'Retry-After': "1"}
+        )
+
+        with pytest.raises(SlackConnectionError) as e:
+            server.rtm_connect()
+            for call in rsps.calls:
+                assert call.response.status_code == 429
+            assert e.message == "RTM connection attempt was rate limited 10 times."
+
+
+def test_custom_agent(server):
+    server.append_user_agent("test agent", 1.0)
+    assert server.api_requester.custom_user_agent[0] == ['test agent', 1.0]
 
 
 def test_server_parse_channel_data(server, rtm_start_fixture):
@@ -34,23 +84,133 @@ def test_server_parse_channel_data(server, rtm_start_fixture):
 def test_server_parse_user_data(server, rtm_start_fixture):
     server.parse_user_data(rtm_start_fixture["users"])
     # Find user by Name
-    userbyname = server.users.find('fakeuser')
-    assert type(userbyname) == User
-    assert userbyname == "fakeuser"
-    assert userbyname != "someotheruser"
+    user_by_name = server.users.find('fakeuser')
+    assert type(user_by_name) == User
+    assert user_by_name == "fakeuser"
+    assert user_by_name != "someotheruser"
     # Find user by ID
-    userbyid = server.users.find('U10CX1234')
-    assert type(userbyid) == User
-    assert userbyid == "fakeuser"
-    assert userbyid.email == 'fakeuser@example.com'
+    user_by_id = server.users.find('U10CX1234')
+    assert type(user_by_id) == User
+    assert user_by_id == "fakeuser"
+    assert user_by_id.email == 'fakeuser@example.com'
     # Don't find invalid user
-    userbyid = server.users.find('invaliduser')
-    assert type(userbyid) != User
+    user_by_id = server.users.find('invaliduser')
+    assert user_by_id is None
 
 
 def test_server_cant_connect(server):
-    with pytest.raises(SlackLoginError):
-        reply = server.ping()
+    with responses.RequestsMock() as rsps:
+        rsps.add(
+            responses.POST,
+            "https://slack.com/api/rtm.start",
+            status=403,
+            json={"ok": False}
+        )
+
+        with pytest.raises(SlackConnectionError) as e:
+            server.rtm_connect()
+
+
+def test_reconnect_flag(server, rtm_start_fixture):
+    with responses.RequestsMock() as rsps:
+        rsps.add(
+            responses.POST,
+            "https://slack.com/api/rtm.start",
+            status=200,
+            json=rtm_start_fixture
+        )
+
+        server.rtm_connect(auto_reconnect=True)
+        assert server.auto_reconnect is True
+
+        for call in rsps.calls:
+            assert call.request.url in [
+                "https://slack.com/api/rtm.start"
+            ]
+
+
+def test_rtm_reconnect(server, rtm_start_fixture):
+    with responses.RequestsMock() as rsps:
+        rsps.add(
+            responses.POST,
+            "https://slack.com/api/rtm.connect",
+            status=200,
+            json=rtm_start_fixture
+        )
+
+        server.rtm_connect(auto_reconnect=True, reconnect=True, use_rtm_start=False)
+
+        for call in rsps.calls:
+            assert call.request.url in [
+                "https://slack.com/api/rtm.connect"
+            ]
+
+
+@patch('time.sleep', return_value=None)
+def test_rtm_max_reconnect_timeout(patched_time_sleep, server, rtm_start_fixture):
+    with responses.RequestsMock() as rsps:
+        rsps.add(
+            responses.POST,
+            "https://slack.com/api/rtm.connect",
+            status=200,
+            json=rtm_start_fixture
+        )
+
+        server.reconnect_count = 4
+        server.last_connected_at = time.time()
+        server.rtm_connect(auto_reconnect=True, reconnect=True, use_rtm_start=False)
+
+        assert server.reconnect_count == 5
+
+
+def test_rtm_reconnect_timeout_recently_connected(server, rtm_start_fixture):
+    # If reconnected recently, server must wait to reconnect and increment the counter
+    with responses.RequestsMock() as rsps:
+        rsps.add(
+            responses.POST,
+            "https://slack.com/api/rtm.connect",
+            status=200,
+            json=rtm_start_fixture
+        )
+
+        server.reconnect_count = 0
+        server.last_connected_at = time.time()
+        server.rtm_connect(auto_reconnect=True, reconnect=True, use_rtm_start=False)
+
+        assert server.reconnect_count == 1
+        for call in rsps.calls:
+            assert call.request.url in [
+                "https://slack.com/api/rtm.connect"
+            ]
+
+
+def test_rtm_reconnect_timeout_not_recently_connected(server, rtm_start_fixture):
+    # If reconnecting after 3 minutes since last reconnect, reset counter and connect without wait
+    with responses.RequestsMock() as rsps:
+        rsps.add(
+            responses.POST,
+            "https://slack.com/api/rtm.connect",
+            status=200,
+            json=rtm_start_fixture
+        )
+
+        server.reconnect_count = 1
+        server.last_connected_at = time.time() - 180
+        server.rtm_connect(auto_reconnect=True, reconnect=True, use_rtm_start=False)
+
+        assert server.reconnect_count == 0
+        for call in rsps.calls:
+            assert call.request.url in [
+                "https://slack.com/api/rtm.connect"
+            ]
+
+
+def test_max_rtm_reconnects(server, monkeypatch):
+    monkeypatch.setattr("time.sleep", None)
+    with pytest.raises(SlackConnectionError) as e:
+        server.reconnect_count = 5
+        server.rtm_connect(auto_reconnect=True, reconnect=True, use_rtm_start=False)
+        assert e.message == "RTM connection failed, reached max reconnects."
 
 
 @pytest.mark.xfail