+++ tests/test_slackrequest.py
@@ -32,6 +32,46 @@ def test_custom_user_agent(mocker):
     assert "baragent:2/0.2" in kwargs['headers']['user-agent']
 
 
+def test_auth_header(mocker):
+    requests = mocker.patch('slackclient.slackrequest.requests')
+    request = SlackRequest()
+
+    request.do('xoxb-123', 'chat.postMessage', {'text': 'test', 'channel': '#general'})
+    args, kwargs = requests.post.call_args
+
+    assert "Bearer xoxb-123" in kwargs['headers']['Authorization']
+
+
+def test_token_override(mocker):
+    requests = mocker.patch('slackclient.slackrequest.requests')
+    request = SlackRequest()
+
+    request.do('xoxb-123', 'chat.postMessage',
+               {
+                   'token': "newtoken",
+                   'text': 'test',
+                   'channel': '#general'
+                })
+    args, kwargs = requests.post.call_args
+
+    assert "Bearer newtoken" in kwargs['headers']['Authorization']
+
+
+def test_plural_field(mocker):
+    requests = mocker.patch('slackclient.slackrequest.requests')
+    request = SlackRequest()
+
+    request.do('xoxb-123','conversations.open', {'users': ['U123', 'U234', 'U345']})
+    args, kwargs = requests.post.call_args
+
+    assert kwargs['data'] == {'users': 'U123,U234,U345'}
+
+    request.do('xoxb-123','conversations.open', {'users': "U123,U234,U345"})
+    args2, kwargs2 = requests.post.call_args
+
+    assert kwargs2['data'] == {'users': 'U123,U234,U345'}
+
+
 def test_post_file(mocker):
     requests = mocker.patch('slackclient.slackrequest.requests')
     request = SlackRequest()
@@ -44,8 +84,7 @@ def test_post_file(mocker):
 
     assert requests.post.call_count == 1
     assert 'https://slack.com/api/files.upload' == args[0]
-    assert {'filename': 'slack_logo.png',
-            'token': 'xoxb-123'} == kwargs['data']
+    assert {'filename': 'slack_logo.png'} == kwargs['data']
     assert kwargs['files'] is not None
 
 
@@ -58,8 +97,7 @@ def test_get_file(mocker):
 
     assert requests.post.call_count == 1
     assert 'https://slack.com/api/files.info' == args[0]
-    assert {'file': "myFavoriteFileID",
-            'token': 'xoxb-123'} == kwargs['data']
+    assert {'file': "myFavoriteFileID"} == kwargs['data']
     assert kwargs['files'] is None
 
 
@@ -74,6 +112,4 @@ def test_post_attachements(mocker):
 
     assert requests.post.call_count == 1
     assert 'https://slack.com/api/chat.postMessage' == args[0]
-    assert {'attachments': json.dumps([{'title': 'hello'}]),
-            'token': 'xoxb-123'} == kwargs['data']
-    assert kwargs['files'] is None
+    assert isinstance(kwargs["data"]["attachments"], str)